+++ src/fmt/maps.rs
@@ -45,7 +45,7 @@ pub(super) fn write(
     match fmt.line {
         Repr::Inline => {
             delim_writer(buf, prefix, suffix, |mut buf| {
-                let rm_trailing = map.len() > 0;
+                let rm_trailing = !map.is_empty();
 
                 buf.push(' ');